+++ DalaMock.Sample/Mock/DalamudMockPluginTest.cs
@@ -15,8 +15,8 @@ public class DalamudMockPluginTest : DalamudPluginTest
 {
     private readonly ILogger<DalamudMockPluginTest> logger;
 
-    public DalamudMockPluginTest(IDalamudPluginInterface pluginInterface, ILogger<DalamudMockPluginTest> logger, IPluginLog pluginLog, IFramework framework, ICommandManager commandManager, IDataManager dataManager, ITextureProvider textureProvider, IChatGui chatGui, IDtrBar dtrBar)
-        : base(pluginInterface, pluginLog, framework, commandManager, dataManager, textureProvider, chatGui, dtrBar)
+    public DalamudMockPluginTest(IDalamudPluginInterface pluginInterface, ILogger<DalamudMockPluginTest> logger, IPluginLog pluginLog, IFramework framework, ICommandManager commandManager, IDataManager dataManager, ITextureProvider textureProvider, IChatGui chatGui, IDtrBar dtrBar, ISigScanner sigScanner)
+        : base(pluginInterface, pluginLog, framework, commandManager, dataManager, textureProvider, chatGui, dtrBar, sigScanner)
     {
         this.logger = logger;
         this.logger.LogInformation("Plugin mock started");